# openapi.yaml
openapi: 3.0.0
info:
  title: User Service API
  description: Sync REST API contracts for Identity, Profile Management, and Administration.
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Local Development Server via API Gateway

paths:
  /auth/register:
    post:
      summary: Register a new user account 
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User account and profile successfully initialized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error or email already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

  /auth/login:
    post:
      summary: Authenticate user and issue JWT tokens
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '200':
          description: Successful authentication, stateless tokens generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Invalid email or password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '403':
          description: Account de-activated by administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

  /profiles/me:
    put:
      summary: Update personal profile data
      operationId: updateProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdate'
      responses:
        '200':
          description: Profile successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          description: Missing or invalid JWT Access Token.
        '403':
          description: Operation not permitted.

  /admin/users/{user_id}/lock:
    post:
      summary: Administrative account lockout/ban
      operationId: lockUser
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          description: The UUID of the user to lockout
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User account successfully deactivated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Invalid administrator credentials.
        '403':
          description: Requires explicit Admin role (RBAC).
        '404':
          description: User entity not found.

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your valid JWT Access Token.

  schemas:
    UserCreate:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          format: email
          example: student@elearning.com
        password:
          type: string
          format: password
          example: SecureP@ss123
          minLength: 8

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 4fbc87a2-19e0-47cb-b09e-71649d33ed86
        username:
          type: string
          format: email
          example: student@elearning.com

    Token:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          example: d9f8e7d6-c5b4-a3f2-e1d0...
        token_type:
          type: string
          default: bearer
          example: bearer

    ProfileUpdate:
      type: object
      properties:
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        avatar_url:
          type: string
          format: uri
          nullable: true
          example: https://s3.amazonaws.com/bucket/avatars/user1.jpg
        bio:
          type: string
          nullable: true
          example: Software engineering enthusiast and online learner.

    ProfileResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 8a7b6c5d-4e3f-2a1b-0c9d-8e7f6a5b4c3d
        user_id:
          type: string
          format: uuid
          example: 4fbc87a2-19e0-47cb-b09e-71649d33ed86
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        avatar_url:
          type: string
          example: https://s3.amazonaws.com/bucket/avatars/user1.jpg
        bio:
          type: string
          example: Software engineering enthusiast and online learner.

    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: Operation completed successfully.